060820-17:07
Removed move() and both moveTo() functions. All PhysicsObjects are now in a state of de facto movement as implemented by their mechanics - movement is now by virtue applied via an application of force (which may be set in the constructor and incremented via keypresses).

Collision and gravity are still bugged.